Cleaned up Sun acquisition
