+++ flight/apps/adcs/sun.py
@@ -14,6 +14,8 @@ both for the mode transitions, sun pointing controller accuracy, and attitude de
 
 """
 
+from apps.adcs.consts import PhysicalConst
+from apps.adcs.math import invert_3x3_psd
 from core import logger
 from hal.configuration import SATELLITE
 from micropython import const
@@ -30,15 +32,7 @@ class SUN_VECTOR_STATUS:
     UNDETERMINED_VECTOR = 0x1  # Vector computed with less than 3 lux readings
     NOT_ENOUGH_READINGS = 0x2  # Computation failed due to lack of readings (less than 3 valid readings)
     NO_READINGS = 0x3
-    MISSING_XP_READING = 0x4
-    MISSING_XM_READING = 0x5
-    MISSING_YP_READING = 0x6
-    MISSING_YM_READING = 0x7
-    MISSING_ZM_READING = 0x8
-    MISSING_FULL_X_AXIS_READING = 0x9
-    MISSING_FULL_Y_AXIS_READING = 0xA
-    MISSING_FULL_Z_AXIS_READING = 0xB
-    ECLIPSE = 0xC
+    ECLIPSE = 0x4
 
 
 def _read_light_sensor(face):
@@ -66,9 +60,6 @@ def read_light_sensors():
             logger.warning(f"Error reading {face}: {e}")
             lux_readings.append(ERROR_LUX)
 
-    # Placeholder for the z+ face pyramid reading
-    # lux_readings.append(ERROR_LUX)
-
     return lux_readings
 
 
@@ -87,50 +78,36 @@ def compute_body_sun_vector_from_lux(I_vec):
     status = None
     sun_body = np.zeros(3)
 
-    N = np.array(
-        [
-            [1, 0, 0],
-            [-1, 0, 0],
-            [0, 1, 0],
-            [0, -1, 0],
-            [0, 0, -1],
-            [0.7071, 0, 0.7071],
-            [0, 0.7071, 0.7071],
-            [-0.7071, 0, 0.7071],
-            [0, -0.7071, 0.7071],
-        ]
-    )  # map from light sensors to body vector
-
+    # Determine Sun Status
     num_valid_readings = NUM_LIGHT_SENSORS - I_vec.count(ERROR_LUX)
-
     if num_valid_readings == 0:
         status = SUN_VECTOR_STATUS.NO_READINGS
         return status, sun_body
     elif num_valid_readings < 3:
         status = SUN_VECTOR_STATUS.NOT_ENOUGH_READINGS
+    elif in_eclipse(I_vec):
+        status = SUN_VECTOR_STATUS.ECLIPSE
+        return status, sun_body
     elif num_valid_readings >= 3:  # All readings are valid and unique determination is possible
         status = SUN_VECTOR_STATUS.UNIQUE_DETERMINATION
 
-    valid_sensor_idxs = [idx for idx in range(len(I_vec)) if I_vec[idx] != ERROR_LUX]
-
     # Extract body vectors and lux readings where the sensor readings are valid
-    N_valid = N[valid_sensor_idxs, :]
+    valid_sensor_idxs = [
+        idx for idx in range(len(I_vec)) if I_vec[idx] != ERROR_LUX and I_vec[idx] >= THRESHOLD_ILLUMINATION_LUX
+    ]
+    N_valid = PhysicalConst.LIGHT_SENSOR_NORMALS[valid_sensor_idxs, :]
     I_valid = [I_vec[idx] for idx in valid_sensor_idxs]
-    
-    # twice faster than linalg inv
-    oprod_sun_inv = invert_3x3_psd(N_valid.transpose() @ N_valid)
-    # Confirm unique determination. If not invertible, not unique
-    if oprod_sun_inv is None:
+
+    # Compute the Inverse of the valid light sensor normals using the Moore-Penrose pseudo-inverse
+    oprod_sun_inv = invert_3x3_psd(np.dot(N_valid.transpose(), N_valid))
+    if oprod_sun_inv is None:  # If the inverse is not possible, sun positioning is not uniquely determinable
         status = SUN_VECTOR_STATUS.NOT_ENOUGH_READINGS
         return status, sun_body
+    else:
+        oprod_sun_inv = np.dot(oprod_sun_inv, N_valid.transpose())
 
     # Extract the sun body vector
-    # NOTE : ulab does not have a pinv operation
-    # Using the Moore-Penrose psuedo-inverse
-    sun_body = oprod_sun_inv @ N_valid.transpose() @ I_valid
-    """
-    sun_body = np.linalg.inv(N_valid.transpose() @ N_valid) @ N_valid.transpose() @ I_valid
-    """
+    sun_body = np.dot(oprod_sun_inv, I_valid)
     norm = (sun_body[0] ** 2 + sun_body[1] ** 2 + sun_body[2] ** 2) ** 0.5
 
     if norm == 0:  # Avoid division by zero - not perfect
@@ -168,11 +145,6 @@ def in_eclipse(raw_readings, threshold_lux_illumination=THRESHOLD_ILLUMINATION_L
     return eclipse
 
 
-def read_pyramid_sun_sensor_zm():
-    # TODO
-    pass
-
-
 def unix_time_to_julian_day(unix_time):
     """Takes in a unix timestamp and returns the julian day"""
     return unix_time / 86400 + 2440587.5
@@ -201,37 +173,3 @@ def approx_sun_position_ECI(utime):
     r_vec = np.array([r_mag * np.cos(long), r_mag * np.sin(long) * np.cos(epsilon), r_mag * np.sin(long) * np.sin(epsilon)])
 
     return r_vec
-
-def invert_3x3_psd(matrix):
-    """
-    Inverts a 3x3 symmetric matrix using the classical adjoint (cofactor) method.
-    Returns a new 3x3 matrix that is the inverse of 'matrix' or none if it is not invertible.
-    About twice faster than np.linalg.inv
-    """
-    if matrix.shape != (3, 3):
-        raise ValueError("Matrix must be 3x3.")
-
-    # Helper aliases for clarity
-    a, b, c = matrix[0]
-    _, e, f = matrix[1]
-    _, _, i = matrix[2]
-
-    # Calculate the determinant
-    det = (a * (e*i - f*f)
-           - b * (b*i - 2*f*c)
-           - e*c*c)
-
-    if abs(det) < 1e-12:  # If determinant is (near) zero, it's not invertible
-        return None
-
-    # Calculate cofactors (matrix of minors with alternating signs)
-    # Cofactor matrix (not yet transposed):
-    adjugate = np.array([
-        [(e*i - f*f),        -(b*i - f*c),        (b*f - e*c)],
-        [-(b*i - c*f),       (a*i - c*c),         -(a*f - b*c)],
-        [(b*f - c*e),        -(a*f - c*b),        (a*e - b*b)]
-    ])
-
-    # Multiply adjugate by 1/det to get the inverse
-    inv = adjugate / det
-    return inv
